pyproject.toml
[project]
name = "mlx-omni-server"
version = "0.4.3"
description = "MLX Omni Server is a server that provides OpenAI-compatible APIs using Apple's MLX framework."
authors = [{ name = "madroid", email = "madroidmaq@gmail.com" }]
requires-python = ">=3.11"
readme = "README.md"
license = "MIT"
keywords = [
    "mlx",
    "ai",
    "agi",
    "aigc",
    "server",
    "openai",
    "tts",
    "stt",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    # core
    "fastapi>=0.115.4,<0.116",
    "pydantic>=2.9.2,<3",
    "uvicorn>=0.34.0,<0.35",
    "rich>=13.9.4",
    # chat
    "mlx-lm>=0.24,<0.25",
    "python-multipart>=0.0.20,<0.0.21",
    "sse-starlette>=2.1.3,<3",
    "outlines>=0.1.11,<0.2",
    # models
    "huggingface-hub>=0.30",
    # audio
    "f5-tts-mlx>=0.2.5,<0.3",
    "mlx-whisper>=0.4.1,<0.5",
    "mlx-audio>=0.2.2",
    "numba>=0.57.0",
    # images
    "mflux>=0.7.1,<0.8",
    # embeddings
    "mlx-embeddings>=0.0.3",
    "pydantic_settings",
    "python-dotenv"
]

[project.urls]
Repository = "https://github.com/madroidmaq/mlx-omni-server"

[project.scripts]
mlx-omni-server = "mlx_omni_server.main:start"

[dependency-groups]
dev = [
    "pytest>=8.3.3,<9",
    "httpx>=0.27.2,<0.28",
    "pre-commit>=4.0.1,<5",
    "black>=24.10.0,<25",
    "isort>=5.13.2,<6",
    "weave>=0.51.46,<0.60",
    "openai>=1.78.0",
    # for en_core_web_sm models
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
]

[tool.hatch.build.targets.sdist]
include = ["src/mlx_omni_server"]

[tool.hatch.build.targets.wheel]
include = ["src/mlx_omni_server"]

[tool.hatch.build.targets.wheel.sources]
"src/mlx_omni_server" = "mlx_omni_server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
